Fixing bench overflow in GFLOPS computation
Fixing a bug in some bench applications where GFLOPS computation ran into integer overflows because explicit type casting to double was not done in the computation

removing all multiplies by 1.0 during GFLOP computation

AMD-Internal: CPUPL-7016

---------

+++ bench/bench_asumv.c
@@ -202,7 +202,7 @@ int main( int argc, char** argv )
             dtime_save = bli_clock_min_diff( dtime_save, dtime );
           }
 
-        gflops = n / ( dtime_save * 1.0e9 );
+        gflops = (double)n / ( dtime_save * 1.0e9 );
         if ( bli_is_complex( dt ) ) gflops *= 2.0;
 
         printf( "data_asumv_%s", BLAS );
