+++ bench/bench_scalv.c
@@ -274,7 +274,7 @@ int main( int argc, char** argv )
             dtime_save = bli_clock_min_diff( dtime_save, dtime );
           }
 
-        gflops = n / ( dtime_save * 1.0e9 );
+        gflops = (double)n / ( dtime_save * 1.0e9 );
 
         if ( bli_is_complex( dt_x ) ) 
         {
